Adding VersionedLayerClient dummy class.
This CR adds new empty class for dataservice-write
and test for initialization.

Adds olp-sdk-core as a dependency. Removes dummy HelloWorld class.

Split integration tests. And use LIB_VERSION in test from core package.

Resolves: OLPEDGE-1935.
