+++ @here/olp-sdk-dataservice-write/lib/client/VersionedLayerClient.ts
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2020 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+import { HRN, OlpClientSettings } from "@here/olp-sdk-core";
+
+/**
+ * Parameters for use to initialize VolatileLayerClient.
+ */
+export interface VersionedLayerClientParams {
+    // [[HRN]] instance of the catalog hrn.
+    catalogHrn: HRN;
+    // The [[OlpClientSettings]] instance.
+    settings: OlpClientSettings;
+}
+
+/**
+ * Describes a versioned layer and provides the possibility to push the data to the versioned layers.
+ */
+export class VersionedLayerClient {
+    /**
+     * Creates the [[VersionedLayerClient]] instance with VersionedLayerClientParams.
+     *
+     * @param params parameters for use to initialize VersionedLayerClient.
+     */
+    constructor(private readonly params: VersionedLayerClientParams) {}
+}
