+++ @here/olp-sdk-dataservice-write/lib/client/index.ts
@@ -17,4 +17,4 @@
  * License-Filename: LICENSE
  */
 
-export class HelloWorld {}
+export * from "./VersionedLayerClient";
