+++ tests/integration/olp-sdk-dataservice-read/Metadata-api.test.ts
@@ -26,7 +26,7 @@ import {
   RequestFactory
 } from "@here/olp-sdk-dataservice-read";
 import { MetadataApi } from "@here/olp-sdk-dataservice-api";
-import { FetchMock } from "./FetchMock";
+import { FetchMock } from "../FetchMock";
 
 chai.use(sinonChai);
 const expect = chai.expect;
