+++ tests/integration/olp-sdk-dataservice-read/VolatileLayerClient.test.ts
@@ -21,31 +21,30 @@ import * as sinon from "sinon";
 import * as chai from "chai";
 import sinonChai = require("sinon-chai");
 import {
-  VersionedLayerClient,
+  VolatileLayerClient,
   OlpClientSettings,
   HRN,
   PartitionsRequest,
   DataRequest,
   quadKeyFromMortonCode,
-  QuadKeyPartitionsRequest,
-  FetchOptions
+  QuadKeyPartitionsRequest
 } from "@here/olp-sdk-dataservice-read";
-import { FetchMock } from "./FetchMock";
+import { FetchMock } from "../FetchMock";
 import { Buffer } from "buffer";
-import { LIB_VERSION } from "@here/olp-sdk-dataservice-read/lib.version";
+import { LIB_VERSION } from "@here/olp-sdk-core";
 
 chai.use(sinonChai);
 
 const assert = chai.assert;
 const expect = chai.expect;
 
-describe("VersionedLayerClient", () => {
+describe("VolatileLayerClient", () => {
   let fetchMock: FetchMock;
   let sandbox: sinon.SinonSandbox;
   let fetchStub: sinon.SinonStub;
 
   const testHRN = HRN.fromString("hrn:here:data:::test-hrn");
-  const testLayerId = "test-layed-id";
+  const testVolatileLayerId = "test-layed-id";
   const headers = new Headers();
   headers.append("cache-control", "max-age=3600");
 
@@ -68,13 +67,13 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
     );
     assert.isDefined(layerClient);
-    expect(layerClient).to.be.instanceOf(VersionedLayerClient);
+    expect(layerClient).to.be.instanceOf(VolatileLayerClient);
   });
 
   it("Shoud be initialization error be handled", async () => {
@@ -83,7 +82,7 @@ describe("VersionedLayerClient", () => {
       getToken: () => Promise.resolve("test-token-string")
     });
     try {
-      const layerClient = new VersionedLayerClient(
+      const layerClient = new VolatileLayerClient(
         HRN.fromString("hrn:here:data:::test-hrn"),
         "",
         settings
@@ -91,43 +90,6 @@ describe("VersionedLayerClient", () => {
     } catch (error) {
       expect(error.message).equal("Unsupported parameters");
     }
-  });
-
-  it("Shoud be initialized with VersionedLayerClientParams with version", async () => {
-    const settings = new OlpClientSettings({
-      environment: "here",
-      getToken: () => Promise.resolve("test-token-string")
-    });
-
-    const versionedLayerClientParams = {
-      catalogHrn: HRN.fromString("hrn:here:data:::test-hrn"),
-      layerId: "test-layed-id",
-      settings: settings,
-      version: 5
-    };
-    const layerClientWithVersion = new VersionedLayerClient(
-      versionedLayerClientParams
-    );
-    assert.isDefined(layerClientWithVersion);
-    expect(layerClientWithVersion).to.be.instanceOf(VersionedLayerClient);
-  });
-
-  it("Shoud be initialised with VersionedLayerClientParams without version", async () => {
-    const settings = new OlpClientSettings({
-      environment: "here",
-      getToken: () => Promise.resolve("test-token-string")
-    });
-
-    const versionedLayerClientParams = {
-      catalogHrn: HRN.fromString("hrn:here:data:::test-hrn"),
-      layerId: "test-layed-id",
-      settings
-    };
-    const layerClientWithoutVersion = new VersionedLayerClient(
-      versionedLayerClientParams
-    );
-    assert.isDefined(layerClientWithoutVersion);
-    expect(layerClientWithoutVersion).to.be.instanceOf(VersionedLayerClient);
   });
 
   it("Shoud be fetched partitions metadata for specific IDs", async () => {
@@ -147,30 +109,15 @@ describe("VersionedLayerClient", () => {
               additionalProp2: "string",
               additionalProp3: "string"
             }
-          },
-          {
-            api: "metadata",
-            version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/metadata/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
           }
         ]),
         { headers }
       )
     );
 
-    mockedResponses.set(
-      `https://query.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
-      new Response(JSON.stringify({ version: 128 }), { headers })
-    );
-
     // Set the response with mocked partitions for IDs 100 and 1000 from Query service
     mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/partitions?partition=100&partition=1000&version=128`,
+      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/partitions?partition=100&partition=1000`,
       new Response(
         JSON.stringify({
           partitions: [
@@ -206,7 +153,7 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
@@ -234,13 +181,94 @@ describe("VersionedLayerClient", () => {
       expect(partitions.partitions[2]).to.be.undefined;
 
       /**
-       * Check if the count of requests are as expected. Should be called 3 times.
+       * Check if the count of requests are as expected. Should be called 2 times.
        * One to the lookup service
-       * for the baseURL to the Query service, one for the latest version,
-       * and last one to the query service.
+       * for the baseURL to the Query service and another one to the query service.
        */
-      expect(fetchStub.callCount).to.be.equal(3);
+      expect(fetchStub.callCount).to.be.equal(2);
     }
+  });
+
+  it("Shoud be fetched data with PartitionId", async () => {
+    const mockedResponses = new Map();
+    const mockedPartitionId = "0000042";
+    const mockedData = Buffer.alloc(42);
+    const mockedPartitionsIdData = {
+      partitions: [
+        {
+          version: 1,
+          partition: "0000042",
+          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
+        },
+        {
+          version: 42,
+          partition: "0000019",
+          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
+        }
+      ]
+    };
+
+    // Set the response from lookup api with the info about Query API.
+    mockedResponses.set(
+      `https://api-lookup.data.api.platform.here.com/lookup/v1/resources/hrn:here:data:::test-hrn/apis`,
+      new Response(
+        JSON.stringify([
+          {
+            api: "query",
+            version: "v1",
+            baseURL: "https://query.data.api.platform.here.com/query/v1",
+            parameters: {
+              additionalProp1: "string",
+              additionalProp2: "string",
+              additionalProp3: "string"
+            }
+          },
+          {
+            api: "volatile-blob",
+            version: "v1",
+            baseURL:
+              "https://volatile-blob.data.api.platform.here.com/volatile-blob/v1",
+            parameters: {
+              additionalProp1: "string",
+              additionalProp2: "string",
+              additionalProp3: "string"
+            }
+          }
+        ]),
+        { headers }
+      )
+    );
+
+    mockedResponses.set(
+      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/partitions?partition=0000042`,
+      new Response(JSON.stringify(mockedPartitionsIdData), { headers })
+    );
+
+    // Set the response of mocked partitions from metadata service.
+    mockedResponses.set(
+      `https://volatile-blob.data.api.platform.here.com/volatile-blob/v1/layers/test-layed-id/data/3C3BE24A341D82321A9BA9075A7EF498.123`,
+      new Response(mockedData, { headers })
+    );
+
+    // Setup the fetch to use mocked responses.
+    fetchMock.withMockedResponses(mockedResponses);
+
+    const settings = new OlpClientSettings({
+      environment: "here",
+      getToken: () => Promise.resolve("test-token-string")
+    });
+    const layerClient = new VolatileLayerClient(
+      testHRN,
+      testVolatileLayerId,
+      settings
+    );
+
+    const request = new DataRequest().withPartitionId(mockedPartitionId);
+
+    const data = await layerClient.getData(request);
+
+    assert.isDefined(data);
+    expect(fetchStub.callCount).to.be.equal(3);
   });
 
   it("Shoud be fetched partitions all metadata", async () => {
@@ -266,15 +294,9 @@ describe("VersionedLayerClient", () => {
       )
     );
 
-    // Set the response from Metadata service with the info about latest catalog version.
-    mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
-      new Response(JSON.stringify({ version: 128 }), { headers })
-    );
-
     // Set the response of mocked partitions from metadata service.
     mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/layers/test-layed-id/partitions?version=128`,
+      `https://metadata.data.api.platform.here.com/metadata/v1/layers/test-layed-id/partitions`,
       new Response(
         JSON.stringify({
           partitions: [
@@ -284,8 +306,7 @@ describe("VersionedLayerClient", () => {
               dataHandle: "1b2ca68f-d4a0-4379-8120-cd025640510c",
               dataSize: 1024,
               crc: "c3f276d7",
-              partition: "314010583",
-              version: 1
+              partition: "314010583"
             },
             {
               checksum: "123f66029c232400e3403cd6e9cfd45b",
@@ -293,26 +314,7 @@ describe("VersionedLayerClient", () => {
               dataHandle: "1b2ca68f-d4a0-4379-8120-cd025640578e",
               dataSize: 2084,
               crc: "c3f2766y",
-              partition: "1000",
-              version: 2
-            },
-            {
-              checksum: "123f66029c232400e3403cd6e9cfd345",
-              compressedDataSize: 2084,
-              dataHandle: "1b2ca68f-d4a0-4379-8120-cd0256405444",
-              dataSize: 2084,
-              crc: "c3f2766y",
-              partition: "1000",
-              version: 2
-            },
-            {
-              checksum: "123f66029c232400e3403cd6e9cfd234",
-              compressedDataSize: 2084,
-              dataHandle: "1b2ca68f-d4a0-4379-8120-cd0256405555",
-              dataSize: 2084,
-              crc: "c3f2766y",
-              partition: "1000",
-              version: 2
+              partition: "1000"
             }
           ],
           next: "/uri/to/next/page"
@@ -329,7 +331,7 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
@@ -339,33 +341,36 @@ describe("VersionedLayerClient", () => {
     const request = new PartitionsRequest();
 
     // Send request for partitions metadata.
-    const partitions = await layerClient.getPartitions(request);
+    const partitions = await layerClient.getPartitions(request).catch(error => {
+      console.log(`Error getting partitions: ${error}`);
+    });
 
     // Check if partitions fetched succesful.
     assert.isDefined(partitions);
 
+    // Check if partitions returns as expected.
     if (partitions) {
-      // Check if partitions returns as expected.
       expect(partitions.partitions[0].dataHandle).to.be.equal(
         "1b2ca68f-d4a0-4379-8120-cd025640510c"
       );
       expect(partitions.partitions[1].dataHandle).to.be.equal(
         "1b2ca68f-d4a0-4379-8120-cd025640578e"
       );
-      expect(partitions.partitions.length).to.be.equal(4);
+      expect(partitions.partitions[0].partition).to.be.equal("314010583");
+      expect(partitions.partitions[1].partition).to.be.equal("1000");
+      expect(partitions.partitions.length).to.be.equal(2);
     }
 
     /**
      * Check if the count of requests are as expected.
-     * Should be called 3 times. One to the lookup service
-     * for the baseURL to the Metadata service, the second
-     * one to the Metadata service for getting the latest version
-     * of catalog and another one to the metadata service for the partitions metadata.
+     * Should be called 2 times. One to the lookup service
+     * for the baseURL to the Metadata service and another one
+     * to the metadata service for the partitions metadata.
      */
-    expect(fetchStub.callCount).to.be.equal(3);
+    expect(fetchStub.callCount).to.be.equal(2);
   });
 
-  it("Shoud be fetched data with dataHandle", async () => {
+  it("Shoud read data with dataHandle", async () => {
     const mockedResponses = new Map();
     const mockedDataHandle = "1b2ca68f-d4a0-4379-8120-cd025640510c";
     const mockedData = Buffer.alloc(42);
@@ -376,19 +381,10 @@ describe("VersionedLayerClient", () => {
       new Response(
         JSON.stringify([
           {
-            api: "blob",
+            api: "volatile-blob",
             version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "metadata",
-            version: "v1",
-            baseURL: "https://metadata.data.api.platform.here.com/metadata/v1",
+            baseURL:
+              "https://volatile-blob.data.api.platform.here.com/volatile-blob/v1",
             parameters: {
               additionalProp1: "string",
               additionalProp2: "string",
@@ -402,7 +398,7 @@ describe("VersionedLayerClient", () => {
 
     // Set the response of mocked partitions from metadata service.
     mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/1b2ca68f-d4a0-4379-8120-cd025640510c`,
+      `https://volatile-blob.data.api.platform.here.com/volatile-blob/v1/layers/test-layed-id/data/1b2ca68f-d4a0-4379-8120-cd025640510c`,
       new Response(mockedData, { headers })
     );
 
@@ -413,9 +409,9 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       testHRN,
-      testLayerId,
+      testVolatileLayerId,
       settings
     );
     const request = new DataRequest().withDataHandle(mockedDataHandle);
@@ -423,216 +419,13 @@ describe("VersionedLayerClient", () => {
     const data = await layerClient.getData(request);
 
     assert.isDefined(data);
-    expect(fetchStub.callCount).to.be.equal(2); // 1 - lookup, 1 - blob
-  });
-
-  it("Shoud be fetched data with PartitionId", async () => {
-    const mockedResponses = new Map();
-    const mockedPartitionId = "0000042";
-    const mockedData = Buffer.alloc(42);
-    const mockedPartitionsIdData = {
-      partitions: [
-        {
-          version: 1,
-          partition: "0000042",
-          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
-        },
-        {
-          version: 42,
-          partition: "0000013",
-          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
-        }
-      ]
-    };
-
-    mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/partitions?partition=0000042&version=42`,
-      new Response(JSON.stringify(mockedPartitionsIdData), { headers })
-    );
-
-    mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
-      new Response(JSON.stringify({ version: 42 }), { headers })
-    );
-
-    // Set the response from lookup api with the info about Metadata service.
-    mockedResponses.set(
-      `https://api-lookup.data.api.platform.here.com/lookup/v1/resources/hrn:here:data:::test-hrn/apis`,
-      new Response(
-        JSON.stringify([
-          {
-            api: "query",
-            version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/query/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "metadata",
-            version: "v1",
-            baseURL: "https://metadata.data.api.platform.here.com/metadata/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "blob",
-            version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          }
-        ]),
-        { headers }
-      )
-    );
-
-    // Set the response of mocked partitions from metadata service.
-    mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/3C3BE24A341D82321A9BA9075A7EF498.123`,
-      new Response(mockedData, { headers })
-    );
-
-    // Setup the fetch to use mocked responses.
-    fetchMock.withMockedResponses(mockedResponses);
-
-    const settings = new OlpClientSettings({
-      environment: "here",
-      getToken: () => Promise.resolve("test-token-string")
-    });
-    const layerClient = new VersionedLayerClient(
-      testHRN,
-      testLayerId,
-      settings
-    );
-    const request = new DataRequest().withPartitionId(mockedPartitionId);
-
-    const data = await layerClient.getData(request);
-
-    assert.isDefined(data);
-    expect(fetchStub.callCount).to.be.equal(4);
-  });
-
-  it("Shoud be fetched data with PartitionId and version", async () => {
-    const mockedResponses = new Map();
-    const mockedPartitionId = "0000042";
-    const mockedData = Buffer.alloc(42);
-    const mockedPartitionsIdData = {
-      partitions: [
-        {
-          version: 1,
-          partition: "0000042",
-          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
-        },
-        {
-          version: 42,
-          partition: "0000013",
-          dataHandle: "3C3BE24A341D82321A9BA9075A7EF498.123"
-        }
-      ]
-    };
-
-    mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/partitions?partition=0000042&version=42`,
-      new Response(JSON.stringify(mockedPartitionsIdData), { headers })
-    );
-
-    // Set the response from lookup api with the info about Metadata service.
-    mockedResponses.set(
-      `https://api-lookup.data.api.platform.here.com/lookup/v1/resources/hrn:here:data:::test-hrn/apis`,
-      new Response(
-        JSON.stringify([
-          {
-            api: "blob",
-            version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "query",
-            version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/query/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          }
-        ]),
-        { headers }
-      )
-    );
-
-    // Set the response of mocked partitions from metadata service.
-    mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/3C3BE24A341D82321A9BA9075A7EF498.123`,
-      new Response(mockedData, { headers })
-    );
-
-    // Setup the fetch to use mocked responses.
-    fetchMock.withMockedResponses(mockedResponses);
-
-    const settings = new OlpClientSettings({
-      environment: "here",
-      getToken: () => Promise.resolve("test-token-string")
-    });
-    const layerClient = new VersionedLayerClient(
-      testHRN,
-      testLayerId,
-      settings
-    );
-    const request = new DataRequest()
-      .withPartitionId(mockedPartitionId)
-      .withVersion(42);
-
-    const data = await layerClient.getData(request);
-
-    assert.isDefined(data);
-    expect(fetchStub.callCount).to.be.equal(3); // 1 - lookup, 1 - query, 1 - blob
+    expect(fetchStub.callCount).to.be.equal(2);
   });
 
   it("Shoud be fetched data with QuadKey", async () => {
     const mockedResponses = new Map();
     const mockedQuadKey = quadKeyFromMortonCode("23618403");
     const mockedData = Buffer.alloc(42);
-    const mockedQuadKeyTreeData = {
-      subQuads: [
-        {
-          version: 12,
-          subQuadKey: "1",
-          dataHandle: "c9116bb9-7d00-44bf-9b26-b4ab4c274665"
-        }
-      ],
-      parentQuads: [
-        {
-          version: 12,
-          partition: "23618403",
-          dataHandle: "da51785a-54b0-40cd-95ac-760f56fe5457"
-        }
-      ]
-    };
-
-    mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/versions/42/quadkeys/23618403/depths/0`,
-      new Response(JSON.stringify(mockedQuadKeyTreeData), { headers })
-    );
-
-    mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
-      new Response(JSON.stringify({ version: 42 }), { headers })
-    );
 
     // Set the response from lookup api with the info about Metadata service.
     mockedResponses.set(
@@ -650,9 +443,9 @@ describe("VersionedLayerClient", () => {
             }
           },
           {
-            api: "blob",
+            api: "query",
             version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
+            baseURL: "https://query.data.api.platform.here.com/query/v1",
             parameters: {
               additionalProp1: "string",
               additionalProp2: "string",
@@ -660,9 +453,10 @@ describe("VersionedLayerClient", () => {
             }
           },
           {
-            api: "query",
+            api: "volatile-blob",
             version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/query/v1",
+            baseURL:
+              "https://volatile-blob.data.api.platform.here.com/volatile-blob/v1",
             parameters: {
               additionalProp1: "string",
               additionalProp2: "string",
@@ -674,22 +468,50 @@ describe("VersionedLayerClient", () => {
       )
     );
 
+    mockedResponses.set(
+      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/quadkeys/23618403/depths/0`,
+      new Response(
+        JSON.stringify({
+          subQuads: [
+            {
+              version: 12,
+              subQuadKey: "1",
+              dataHandle: "c9116bb9-7d00-44bf-9b26-b4ab4c274665"
+            }
+          ],
+          parentQuads: [
+            {
+              version: 12,
+              partition: "23618403",
+              dataHandle: "da51785a-54b0-40cd-95ac-760f56fe5457"
+            }
+          ]
+        }),
+        { headers }
+      )
+    );
+
     // Set the response of mocked partitions from metadata service.
     mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/c9116bb9-7d00-44bf-9b26-b4ab4c274665`,
+      `https://volatile-blob.data.api.platform.here.com/volatile-blob/v1/layers/test-layed-id/data/c9116bb9-7d00-44bf-9b26-b4ab4c274665`,
       new Response(mockedData, { headers })
     );
 
     // Setup the fetch to use mocked responses.
     fetchMock.withMockedResponses(mockedResponses);
 
+    mockedResponses.set(
+      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
+      new Response(JSON.stringify({ version: 124 }), { headers })
+    );
+
     const settings = new OlpClientSettings({
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       testHRN,
-      testLayerId,
+      testVolatileLayerId,
       settings
     );
     const request = new DataRequest().withQuadKey(mockedQuadKey);
@@ -700,95 +522,10 @@ describe("VersionedLayerClient", () => {
     expect(fetchStub.callCount).to.be.equal(4);
   });
 
-  it("Shoud be fetched data with QuadKey and Version", async () => {
-    const mockedResponses = new Map();
-    const mockedQuadKey = quadKeyFromMortonCode("23618403");
-    const mockedData = Buffer.alloc(42);
-    const mockedQuadKeyTreeData = {
-      subQuads: [
-        {
-          version: 12,
-          subQuadKey: "1",
-          dataHandle: "c9116bb9-7d00-44bf-9b26-b4ab4c274665"
-        }
-      ],
-      parentQuads: [
-        {
-          version: 12,
-          partition: "23618403",
-          dataHandle: "da51785a-54b0-40cd-95ac-760f56fe5457"
-        }
-      ]
-    };
-
-    mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/versions/42/quadkeys/23618403/depths/0`,
-      new Response(JSON.stringify(mockedQuadKeyTreeData), { headers })
-    );
-
-    // Set the response from lookup api with the info about Metadata service.
-    mockedResponses.set(
-      `https://api-lookup.data.api.platform.here.com/lookup/v1/resources/hrn:here:data:::test-hrn/apis`,
-      new Response(
-        JSON.stringify([
-          {
-            api: "blob",
-            version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "query",
-            version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/query/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          }
-        ]),
-        { headers }
-      )
-    );
-
-    // Set the response of mocked partitions from metadata service.
-    mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/c9116bb9-7d00-44bf-9b26-b4ab4c274665`,
-      new Response(mockedData, { headers })
-    );
-
-    // Setup the fetch to use mocked responses.
-    fetchMock.withMockedResponses(mockedResponses);
-
-    const settings = new OlpClientSettings({
-      environment: "here",
-      getToken: () => Promise.resolve("test-token-string")
-    });
-    const layerClient = new VersionedLayerClient(
-      testHRN,
-      testLayerId,
-      settings
-    );
-    const request = new DataRequest()
-      .withQuadKey(mockedQuadKey)
-      .withVersion(42);
-
-    const data = await layerClient.getData(request);
-
-    assert.isDefined(data);
-    expect(fetchStub.callCount).to.be.equal(3); // 1 - lookup, 1 - query, 1 - blob
-  });
-
-  it("Shoud read partitions metadata by QuadKey for specific VersionLayer", async () => {
+  it("Shoud read partitions metadata by QuadKey for specific VolatileLayer", async () => {
     const mockedResponses = new Map();
 
     const billingTag = "billingTag";
-    const mockedVersion = 42;
     const mockedDepth = 3;
     const mockedQuadKey = {
       row: 1,
@@ -796,21 +533,11 @@ describe("VersionedLayerClient", () => {
       level: 3
     };
 
-    // Set the response from lookup api with the info about Query API.
+    // Set the response from lookup api with the info about Metadata service.
     mockedResponses.set(
       `https://api-lookup.data.api.platform.here.com/lookup/v1/resources/hrn:here:data:::test-hrn/apis`,
       new Response(
         JSON.stringify([
-          {
-            api: "query",
-            version: "v1",
-            baseURL: "https://query.data.api.platform.here.com/query/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
           {
             api: "metadata",
             version: "v1",
@@ -820,15 +547,30 @@ describe("VersionedLayerClient", () => {
               additionalProp2: "string",
               additionalProp3: "string"
             }
+          },
+          {
+            api: "query",
+            version: "v1",
+            baseURL: "https://query.data.api.platform.here.com/query/v1",
+            parameters: {
+              additionalProp1: "string",
+              additionalProp2: "string",
+              additionalProp3: "string"
+            }
           }
         ]),
         { headers }
       )
     );
 
-    // Set the response with mocked partitions
     mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/versions/42/quadkeys/70/depths/3`,
+      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1`,
+      new Response(JSON.stringify({ version: 30 }), { headers })
+    );
+
+    // Set the response with mocked partitions for volatile layer
+    mockedResponses.set(
+      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/quadkeys/70/depths/3`,
       new Response(
         JSON.stringify({
           parentQuads: [
@@ -860,11 +602,6 @@ describe("VersionedLayerClient", () => {
       )
     );
 
-    mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1&billingTag=billingTag`,
-      new Response(JSON.stringify({ version: 42 }), { headers })
-    );
-
     // Setup the fetch to use mocked responses.
     fetchMock.withMockedResponses(mockedResponses);
 
@@ -873,7 +610,7 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
@@ -883,9 +620,6 @@ describe("VersionedLayerClient", () => {
     assert.isDefined(quadKeyPartitionsRequest);
     expect(quadKeyPartitionsRequest).be.instanceOf(QuadKeyPartitionsRequest);
 
-    const quadKeyPartitionsRequestWithVersion = quadKeyPartitionsRequest.withVersion(
-      mockedVersion
-    );
     const quadKeyPartitionsRequestWithDepth = quadKeyPartitionsRequest.withDepth(
       mockedDepth
     );
@@ -896,9 +630,6 @@ describe("VersionedLayerClient", () => {
       billingTag
     );
 
-    expect(quadKeyPartitionsRequestWithVersion.getVersion()).to.be.equal(
-      mockedVersion
-    );
     expect(quadKeyPartitionsRequestWithDepth.getDepth()).to.be.equal(
       mockedDepth
     );
@@ -912,7 +643,6 @@ describe("VersionedLayerClient", () => {
     const partitions = await layerClient.getPartitions(
       quadKeyPartitionsRequest
     );
-
     if (partitions.parentQuads) {
       expect(partitions.parentQuads[0].partition).to.be.equal("73982");
     }
@@ -963,15 +693,9 @@ describe("VersionedLayerClient", () => {
       )
     );
 
-    // Set the response from Metadata service with the info about latest catalog version.
-    mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/versions/latest?startVersion=-1&billingTag=billing-tag`,
-      new Response(JSON.stringify({ version: 30 }), { headers })
-    );
-
     // Set the response of mocked partitions with additional fields.
     mockedResponses.set(
-      `https://metadata.data.api.platform.here.com/metadata/v1/layers/test-layed-id/partitions?version=30&additionalFields=dataSize,checksum,compressedDataSize&billingTag=billing-tag`,
+      `https://metadata.data.api.platform.here.com/metadata/v1/layers/test-layed-id/partitions?additionalFields=dataSize,checksum,compressedDataSize`,
       new Response(JSON.stringify(mockedPartitions), { headers })
     );
 
@@ -983,16 +707,15 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
     );
 
-    const requestWithAdditionalFields = new PartitionsRequest()
-      .withAdditionalFields(["dataSize", "checksum", "compressedDataSize"])
-      .withFetchOption(FetchOptions.OnlineIfNotFound)
-      .withBillingTag("billing-tag");
+    const requestWithAdditionalFields = new PartitionsRequest().withAdditionalFields(
+      ["dataSize", "checksum", "compressedDataSize"]
+    );
 
     const partitions = await layerClient.getPartitions(
       requestWithAdditionalFields
@@ -1077,7 +800,7 @@ describe("VersionedLayerClient", () => {
 
     // Set the response of mocked partitions with additional fields.
     mockedResponses.set(
-      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/versions/30/quadkeys/70/depths/0?additionalFields=dataSize,checksum,compressedDataSize`,
+      `https://query.data.api.platform.here.com/query/v1/layers/test-layed-id/quadkeys/70/depths/0?additionalFields=dataSize,checksum,compressedDataSize`,
       new Response(JSON.stringify(mockedPartitions), { headers })
     );
 
@@ -1089,7 +812,7 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       HRN.fromString("hrn:here:data:::test-hrn"),
       "test-layed-id",
       settings
@@ -1113,6 +836,26 @@ describe("VersionedLayerClient", () => {
     assert.isDefined(partitions);
   });
 
+  it("Shoud be initialized with VolatileLayerClientParams", async () => {
+    const settings = new OlpClientSettings({
+      environment: "here",
+      getToken: () => Promise.resolve("test-token-string")
+    });
+
+    const volatileLayerClientParams = {
+      catalogHrn: HRN.fromString("hrn:here:data:::test-hrn"),
+      layerId: "test-layed-id",
+      settings: settings
+    };
+    const volatileLayerClient = new VolatileLayerClient(
+      volatileLayerClientParams
+    );
+
+    assert.isDefined(volatileLayerClient);
+    expect(volatileLayerClient).to.be.instanceOf(VolatileLayerClient);
+    assert.equal(volatileLayerClient["hrn"], "hrn:here:data:::test-hrn");
+  });
+
   it("Should user-agent be added to the each request", async () => {
     const mockedResponses = new Map();
     const mockedDataHandle = "1b2ca68f-d4a0-4379-8120-cd025640510c";
@@ -1124,19 +867,10 @@ describe("VersionedLayerClient", () => {
       new Response(
         JSON.stringify([
           {
-            api: "blob",
+            api: "volatile-blob",
             version: "v1",
-            baseURL: "https://blob.data.api.platform.here.com/blob/v1",
-            parameters: {
-              additionalProp1: "string",
-              additionalProp2: "string",
-              additionalProp3: "string"
-            }
-          },
-          {
-            api: "metadata",
-            version: "v1",
-            baseURL: "https://metadata.data.api.platform.here.com/metadata/v1",
+            baseURL:
+              "https://volatile-blob.data.api.platform.here.com/volatile-blob/v1",
             parameters: {
               additionalProp1: "string",
               additionalProp2: "string",
@@ -1150,7 +884,7 @@ describe("VersionedLayerClient", () => {
 
     // Set the response of mocked partitions from metadata service.
     mockedResponses.set(
-      `https://blob.data.api.platform.here.com/blob/v1/layers/test-layed-id/data/1b2ca68f-d4a0-4379-8120-cd025640510c`,
+      `https://volatile-blob.data.api.platform.here.com/volatile-blob/v1/layers/test-layed-id/data/1b2ca68f-d4a0-4379-8120-cd025640510c`,
       new Response(mockedData, { headers })
     );
 
@@ -1161,9 +895,9 @@ describe("VersionedLayerClient", () => {
       environment: "here",
       getToken: () => Promise.resolve("test-token-string")
     });
-    const layerClient = new VersionedLayerClient(
+    const layerClient = new VolatileLayerClient(
       testHRN,
-      testLayerId,
+      testVolatileLayerId,
       settings
     );
     const request = new DataRequest().withDataHandle(mockedDataHandle);
@@ -1171,7 +905,7 @@ describe("VersionedLayerClient", () => {
     const data = await layerClient.getData(request);
 
     assert.isDefined(data);
-    expect(fetchStub.callCount).to.be.equal(2); // 1 - lookup, 1 - blob
+    expect(fetchStub.callCount).to.be.equal(2);
     const calls = fetchStub.getCalls();
     calls.forEach(call => {
       const callHeaders = call.args[1].headers;
